feat: Add DownloadableFile as domain model
