+++ src/main/kotlin/dev/minuk/opampcommander/application/services/OpampService.kt
@@ -361,6 +361,13 @@ class OpampService(
                 .newBuilder()
                 .setVersion(version)
                 .setHash(ByteString.copyFrom(hash))
+                .setFile(Opamp.DownloadableFile.newBuilder()
+                    .setHeaders(file.headers.toProto())
+                    .setDownloadUrl(file.downloadUrl)
+                    .setContentHash(ByteString.copyFrom(file.contentHash))
+                    .setSignature(ByteString.copyFrom(file.signature))
+                    .build()
+                )
         return builder.build()
     }
