+++ src/main/kotlin/dev/minuk/opampcommander/domain/models/agent/AgentRemoteConfig.kt
@@ -1,0 +1,6 @@
+package dev.minuk.opampcommander.domain.models.agent
+
+data class AgentRemoteConfig(
+    val config: AgentConfigMap,
+    val configHash: ByteArray,
+)
